pyproject.toml
[tool.poetry]
name = "core-tutorials"
version = "0.1.0"
description = ""
authors = ["nikhil <nikhil.bhardwaj@sveltetech.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.8"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
psycopg2 = "^2.9.9"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"
pytest = "^8.2.1"
pytest-django = "^4.8.0"
model-bakery = "^1.18.0"
pytest-xdist = "^3.6.1"
djangorestframework-simplejwt = "^5.3.1"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.pytest.ini_options]
pythonpath = "."
DJANGO_SETTINGS_MODULE = "cooking_core.project.settings"
python_files = "test_*.py"
testpaths = ["cooking_core"]
django_find_project = false
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
